Move legacy notify to legacy
